mysql db 연결 및 간단한 회원관리 기능 구현

+++ Spring-Side/src/main/java/firstSpringWebProject/firstSpringWebProject/domain/Board.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
-import org.springframework.data.annotation.Id;
+
 
 import javax.persistence.*;
 import java.sql.Timestamp;
@@ -22,7 +22,7 @@ public class Board {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false,length = 100)
     private String title;
