+++ Spring-Side/src/main/java/firstSpringWebProject/firstSpringWebProject/domain/User.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.CreationTimestamp;
 
 import javax.persistence.*;
@@ -17,7 +18,7 @@ import java.sql.Timestamp;
 public class User {
     @Id//primarykey 등록
     @GeneratedValue(strategy = GenerationType.IDENTITY)//연결된 DB의 너버링 전략을 따라간다는 설정
-    private int id;//auto_increment
+    private Long id;//auto_increment
 
     @Column(nullable = false,length = 100)//hash 암호화시를 위해 넉넉하게 설정
     private String password;
